refactor: rename of max length number input component to index
