+++ src/components/PaymentModal/AddCardStep/AddCardStep.tsx
@@ -4,7 +4,7 @@ import { CardDateInput } from '@/components/common/CardDateInput'
 import { CardNumberInput } from '@/components/common/CardNumberInput'
 import { CardPasswordInput } from '@/components/common/CardPasswordInput'
 import { Input } from '@/components/common/Input'
-import { MaxLengthNumberInput } from '@/components/common/MaxLengthNumberInput/MaxLengthNumberInput'
+import { MaxLengthNumberInput } from '@/components/common/MaxLengthNumberInput'
 import { useDialog } from '@/hooks/useDialog'
 
 export const AddCardStep = () => {
