+++ src/components/common/CardDateInput/index.tsx
@@ -1,7 +1,7 @@
 import { ChangeEvent, useRef, useState } from 'react'
 import { css } from '@emotion/css'
 import { CARD_DATE } from '@/constants/app'
-import { MaxLengthNumberInput } from '../MaxLengthNumberInput/MaxLengthNumberInput'
+import { MaxLengthNumberInput } from '../MaxLengthNumberInput'
 
 type CardDateValues = {
   'card_date-month': string
